requirements.txt
# ── Core LLM & LangChain ──────────────────────────────────────────────────────
langchain==0.1.20
langchain-community==0.0.38
langchain-huggingface==0.0.3
langchain-groq==0.1.3          # Groq LLM integration for LangChain
langchain-text-splitters==0.0.2

# ── LLM Providers & Embeddings ────────────────────────────────────────────────
groq==0.9.0                    # Groq Python SDK (used by langchain-groq)
huggingface_hub==0.24.6
transformers==4.41.2
sentence-transformers==2.7.0   # MiniLM-L6-v2 embedding model

# ── Vector Store ──────────────────────────────────────────────────────────────
faiss-cpu==1.8.0               # Local FAISS vector store (CPU version)

# ── Document Loaders ──────────────────────────────────────────────────────────
pypdf==4.2.0                   # PDF loading and parsing

# ── Web UI ────────────────────────────────────────────────────────────────────
streamlit==1.35.0              # Streamlit UI framework

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1           # Load secrets from .env file
tqdm==4.66.4
numpy==1.24.4
torch==2.1.0                   # Required by sentence-transformers
